backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-dotenv
requests
python-multipart
# google-generativeai

